ourblog/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cloudinary==1.26.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap5==2.1.2
django-classy-tags==2.0.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-etc==1.3.0
django-extensions==3.1.3
django-filter==2.4.0
django-formtools==2.3
django-mptt==0.13.4
django-notifications-hq==1.6.0
django-sekizai==2.0.0
django-social-share==2.2.1
django-summernote==0.8.11.6
django-taggit==1.5.1
django-treebeard==4.5.1
django-webpush==0.3.3
djangocms-admin-style==2.0.2
djangorestframework==3.12.4
flake8==3.9.2
flake8-isort==4.0.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
http-ece==1.1.0
idna==3.2
importlib-metadata==2.1.1
isort==5.9.3
jsonfield==3.1.0
mccabe==0.6.1
oauthlib==3.1.1
path==16.2.0
pathlib==1.0.1
Pillow==8.3.1
psycopg2==2.9.1
py-vapid==1.8.2
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pymemcache==3.5.0
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
python-http-client==3.3.3
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
pywebpush==1.9.4
reportlab==3.6.1
requests==2.26.0
requests-oauthlib==1.3.0
sendgrid==6.8.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
swapper==1.1.2.post1
testfixtures==6.18.3
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
zipp==3.5.0
